refactor: fix bags, add protectedroutes
