+++ react-burger/src/components/App/App.jsx
@@ -1,33 +1,41 @@
 import { React } from "react";
 import Styles from "./App.module.css";
 import AppHeader from "../AppHeader/AppHeader";
-import BurgerConstructor from "../BurgerConstructor/BurgerConstructor";
-import BurgerIngredients from "../BurgerIngredients/BurgerIngredients";
+import BurgerConstructor from "../../pages/BurgerConstructor/BurgerConstructor";
+import BurgerIngredients from "../../pages/BurgerIngredients/BurgerIngredients";
 import { useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
+import { useDispatch } from "react-redux";
 import { getItems } from "../../services/actions";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
-import Profile from "../Profile/Profile";
-import NotFound from "../NotFound/NotFound";
-import { Route, Switch } from "react-router-dom";
-import Login from "../Login/Login";
-import Register from "../Register/Register";
-import ResetPassword from "../ResetPassword/ResetPassword";
-import ForgotPassword from "../ForgotPassword/ForgotPassword";
+import Profile from "../../pages/Profile/Profile";
+import NotFound from "../../pages/NotFound/NotFound";
+import { Route, Switch, useHistory } from "react-router-dom";
+import Login from "../../pages/Login/Login";
+import Register from "../../pages/Register/Register";
+import ResetPassword from "../../pages/ResetPassword/ResetPassword";
+import ForgotPassword from "../../pages/ForgotPassword/ForgotPassword";
 import IngredientDetails from "../IngredientDetails/IngredientDetails";
 import Modal from "../Modal/Modal";
 import { useLocation } from "react-router-dom";
 import ProtectedRoute from "../ProtectedRoute";
-
+import { checkAuth } from "../../services/actions/auth";
 function App() {
-  let location = useLocation();
-  let background = location.state && location.state.background;
+  const history = useHistory();
+  const back = (e) => {
+    e.stopPropagation();
+    history.goBack();
+  };
+
+  const location = useLocation();
+  const background = location.state && location.state.background;
   const dispatch = useDispatch();
+
   useEffect(() => {
     dispatch(getItems());
-  }, [dispatch]);
-console.log('app')
+    checkAuth(dispatch);
+  }, []);
+
   return (
     <div className={Styles.App} id="app">
       <AppHeader />
@@ -42,23 +50,21 @@ console.log('app')
             </DndProvider>
           </div>
         </Route>
-        <Route path="/login" exact={true}>
+        <ProtectedRoute onlyForAuth={true} path="/login" exact={true}>
           <Login />
-        </Route>
-        <Route path="/register" exact={true}>
+        </ProtectedRoute>
+        <ProtectedRoute onlyForAuth={true} path="/register" exact={true}>
           <Register />
-        </Route>
-        <Route path="/reset-password" exact={true}>
+        </ProtectedRoute>
+        <ProtectedRoute onlyForAuth={true} path="/reset-password" exact={true}>
           <ResetPassword />
-        </Route>
-        <Route path="/profile">
-        <ProtectedRoute path='/login'>
+        </ProtectedRoute>
+        <ProtectedRoute path="/profile">
           <Profile />
-          </ProtectedRoute>
-        </Route>
-        <Route path="/forgot-password" exact={true}>
+        </ProtectedRoute>
+        <ProtectedRoute onlyForAuth={true} path="/forgot-password" exact={true}>
           <ForgotPassword />
-        </Route>
+        </ProtectedRoute>
         <Route path="/ingredient/:id" exact={true}>
           <IngredientDetails />
         </Route>
@@ -69,7 +75,7 @@ console.log('app')
       {background && (
         <Route
           path="/ingredient/:id"
-          children={<Modal children={<IngredientDetails />} />}
+          children={<Modal children={<IngredientDetails />} onClose={back} />}
         />
       )}
     </div>
