+++ react-burger/src/pages/NotFound/NotFound.jsx
@@ -1,0 +1,10 @@
+import styles from "./NotFound.module.css";
+const NotFound = () => {
+  return (
+    <div className={styles.notFound}>
+      <p className="text text_type_main-large">404 NOT FOUND!</p>
+    </div>
+  );
+};
+
+export default NotFound;
