+++ react-burger/src/pages/ResetPassword/ResetPassword.module.css
@@ -7,3 +7,11 @@
     align-items: center;
   gap: 24px;
 }
+.formContainer{
+  width: 480px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+gap: 24px;
+}
